CMake: Add testing support
The examples were used as testing because they are already used to test
stuff. Everything has been moved to the "tests" directory.

Use "cmake --build . --target test" to invoke CTest, that just runs
programs that actually do the tests and report them, if it is
successful or not. Use the target "check" to get verbose output.