mk: List suffixes in one place (to fix openwatcom WMAKE)
Implicit rules in openwatcom WMAKE are weird. It requires both file
extensions to be in `.EXTENSIONS` (alias for `.SUFFIXES`), but the
source file extension needs to be AFTER the destination file
extension, which isn't possible in the previous setup.

On the plus side, it saves some bytes in the makefiles.