mk/watcom: Add workaround for openwatcom include dir search order
The openwatcom compiler prefers to search the current directory
for include files, which isn't optimal for include files that are
named the same (e. g. `wlab/defines.h` and `defines.h`).

For now, "redirect" the top-level header files to the correct
header file by examining what application is currently being build,
which is set in each source file.

This probably shouldn't be long-term. Maybe it's better to `cd` into
the directory before compiling.